Refactor `RAP` and `EliminateBC` in `FABilinearFormExtension`
