+++ fem/lor/lor_batched.cpp
@@ -13,6 +13,7 @@
 #include "../../fem/quadinterpolator.hpp"
 #include "../../general/forall.hpp"
 #include <climits>
+#include "../pbilinearform.hpp"
 
 // Specializations
 #include "lor_h1.hpp"
@@ -511,171 +512,8 @@ void BatchedLORAssembly::ParAssemble(
 
    // Eliminate the boundary conditions
    HypreParMatrix *A_mat = A.As<HypreParMatrix>();
-   hypre_ParCSRMatrix *A_hypre = *A_mat;
-   A_mat->HypreReadWrite();
 
-   hypre_CSRMatrix *diag = hypre_ParCSRMatrixDiag(A_hypre);
-   hypre_CSRMatrix *offd = hypre_ParCSRMatrixOffd(A_hypre);
-
-   HYPRE_Int diag_nrows = hypre_CSRMatrixNumRows(diag);
-   HYPRE_Int offd_ncols = hypre_CSRMatrixNumCols(offd);
-
-   const int n_ess_dofs = ess_dofs.Size();
-   const auto ess_dofs_d = ess_dofs.GetMemory().Read(
-                              GetHypreMemoryClass(), n_ess_dofs);
-
-   // Start communication to figure out which columns need to be eliminated in
-   // the off-diagonal block
-   hypre_ParCSRCommHandle *comm_handle;
-   HYPRE_Int *int_buf_data, *eliminate_row, *eliminate_col;
-   {
-      eliminate_row = mfem_hypre_CTAlloc_host(HYPRE_Int, diag_nrows);
-      eliminate_col = mfem_hypre_CTAlloc_host(HYPRE_Int, offd_ncols);
-
-      // Make sure A has a communication package
-      hypre_ParCSRCommPkg *comm_pkg = hypre_ParCSRMatrixCommPkg(A_hypre);
-      if (!comm_pkg)
-      {
-         hypre_MatvecCommPkgCreate(A_hypre);
-         comm_pkg = hypre_ParCSRMatrixCommPkg(A_hypre);
-      }
-
-      // Which of the local rows are to be eliminated?
-      for (int i = 0; i < diag_nrows; i++)
-      {
-         eliminate_row[i] = 0;
-      }
-
-      ess_dofs.HostRead();
-      for (int i = 0; i < n_ess_dofs; i++)
-      {
-         eliminate_row[ess_dofs[i]] = 1;
-      }
-
-      // Use a matvec communication pattern to find (in eliminate_col) which of
-      // the local offd columns are to be eliminated
-      HYPRE_Int num_sends = hypre_ParCSRCommPkgNumSends(comm_pkg);
-      int_buf_data = mfem_hypre_CTAlloc_host(
-                        HYPRE_Int,
-                        hypre_ParCSRCommPkgSendMapStart(comm_pkg, num_sends));
-      int index = 0;
-      for (int i = 0; i < num_sends; i++)
-      {
-         int start = hypre_ParCSRCommPkgSendMapStart(comm_pkg, i);
-         for (int j = start; j < hypre_ParCSRCommPkgSendMapStart(comm_pkg, i+1); j++)
-         {
-            int k = hypre_ParCSRCommPkgSendMapElmt(comm_pkg,j);
-            int_buf_data[index++] = eliminate_row[k];
-         }
-      }
-      comm_handle = hypre_ParCSRCommHandleCreate(
-                       11, comm_pkg, int_buf_data, eliminate_col);
-   }
-
-   // Eliminate rows and columns in the diagonal block
-   {
-      const auto I = diag->i;
-      const auto J = diag->j;
-      auto data = diag->data;
-
-      MFEM_HYPRE_FORALL(i, n_ess_dofs,
-      {
-         const int idof = ess_dofs_d[i];
-         for (int j=I[idof]; j<I[idof+1]; ++j)
-         {
-            const int jdof = J[j];
-            if (jdof == idof)
-            {
-               // Set eliminate diagonal equal to identity
-               data[j] = 1.0;
-            }
-            else
-            {
-               data[j] = 0.0;
-               for (int k=I[jdof]; k<I[jdof+1]; ++k)
-               {
-                  if (J[k] == idof)
-                  {
-                     data[k] = 0.0;
-                     break;
-                  }
-               }
-            }
-         }
-      });
-   }
-
-   // Eliminate rows in the off-diagonal block
-   {
-      const auto I = offd->i;
-      auto data = offd->data;
-      MFEM_HYPRE_FORALL(i, n_ess_dofs,
-      {
-         const int idof = ess_dofs_d[i];
-         for (int j=I[idof]; j<I[idof+1]; ++j)
-         {
-            data[j] = 0.0;
-         }
-      });
-   }
-
-   // Wait for MPI communication to finish
-   Array<HYPRE_Int> cols_to_eliminate;
-   {
-      hypre_ParCSRCommHandleDestroy(comm_handle);
-
-      // set the array cols_to_eliminate
-      int ncols_to_eliminate = 0;
-      for (int i = 0; i < offd_ncols; i++)
-      {
-         if (eliminate_col[i]) { ncols_to_eliminate++; }
-      }
-
-      cols_to_eliminate.SetSize(ncols_to_eliminate);
-      cols_to_eliminate = 0.0;
-
-      ncols_to_eliminate = 0;
-      for (int i = 0; i < offd_ncols; i++)
-      {
-         if (eliminate_col[i])
-         {
-            cols_to_eliminate[ncols_to_eliminate++] = i;
-         }
-      }
-
-      mfem_hypre_TFree_host(int_buf_data);
-      mfem_hypre_TFree_host(eliminate_row);
-      mfem_hypre_TFree_host(eliminate_col);
-   }
-
-   // Eliminate columns in the off-diagonal block
-   {
-      const int ncols_to_eliminate = cols_to_eliminate.Size();
-      const int nrows_offd = hypre_CSRMatrixNumRows(offd);
-      const auto cols = cols_to_eliminate.GetMemory().Read(
-                           GetHypreMemoryClass(), ncols_to_eliminate);
-      const auto I = offd->i;
-      const auto J = offd->j;
-      auto data = offd->data;
-      // Note: could also try a different strategy, looping over nnz in the
-      // matrix and then doing a binary search in ncols_to_eliminate to see if
-      // the column should be eliminated.
-      MFEM_HYPRE_FORALL(idx, ncols_to_eliminate,
-      {
-         const int j = cols[idx];
-         for (int i=0; i<nrows_offd; ++i)
-         {
-            for (int jj=I[i]; jj<I[i+1]; ++jj)
-            {
-               if (J[jj] == j)
-               {
-                  data[jj] = 0.0;
-                  break;
-               }
-            }
-         }
-      });
-   }
+   ParBilinearForm::ParallelEliminateBC(ess_dofs, *A_mat);
 }
 #endif
